Don't use AMQP for publishing and consuming
`amqp` prior to version 1.0 didn't have a good error handling,
for example, it would throw a MatchError if an operation, such
as declaring an exchange, was performed on a channel in state `:closing`.

Error handling was added in version `1.0`, but upgrading to `1.0`
because of that doesn't seem very compelling as it forces you to
upgrade `amqp_client` to version `3.7`, which, in turn, brings bunch
of unwanted and maybe conflicting dependencies (such as `ranch 1.5`,
which conflicts with all projects that use `cowboy` 1.x).

This commit basically inlines part of `amqp` library into `freddy`
with few optimizations and proper error handling. We still use `amqp`
to establish connection, open a channel + undocumented `AMQP.Core` that
imports record definitions from `amqp_client` itself.

Another problem that is fixed with this inlining is a noise that pollutes
logs when application shuts down. What happens with `amqp` is that when
you register your process as a consumer using `amqp`, it instead spawns
a hidden "adapter" process and registers adapter as a consumer. Adapter
accepts messages and sends them to your real consumer and also sends
"cancel" to AMQP server when your process dies for any reason. This may
happen when application shuts down too. In this case AMQP server will send
back message "cancel_ok", but there will be nobody to consume this message
which causes `amqp_client` to log some errors.

Now, because we don't use most of the `amqp`, we can relax this dependency
and let library users decide, which version of `amqp` and `amqp_client`
they prefer.