Introduce AMQP connection adapter
This change brings back concept of RabbitMQ adapters. Primary
motivation for this change is that using real RabbitMQ connection
in unit tests is not viable or easy, because real messages may
interfere unit tests and make test suite extermely unstable.

So we want to bring back Sandox adapter, that was removed
along with Hare in 0.12.0.

This commit introduces adapter interface and extracts all
AMQP related operators into adapter.