Fix failing CI
CI was failing with OTP 23.0 because the repository where OTP is pulled
from does not build OTP 23.0 for latest Ubuntu (22.04).

Here is a list of available OTP versions for Ubuntu 22.04:
https://repo.hex.pm/builds/otp/ubuntu-22.04/builds.txt

This commit sets the GitHub runner to Ubuntu 20.04 to fix the build.

Here is a list of available OTP versions for Ubuntu 20.04:
https://repo.hex.pm/builds/otp/ubuntu-20.04/builds.txt

AUTO-657