Update GitHub actions/cache action
The previous version is no longer working and is failing with this
error:

```
Error: This request has been automatically failed because it uses a deprecated version of `actions/cache: v2`. Please update your workflow to use v3/v4 of actions/cache to avoid interruptions. Learn more: https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down
```

On the README page at https://github.com/actions/cache there is no
mention of any breaking changes.

AUTO-1830